+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [4.3.0] - 2021-02-03
 ### Changed
 - Upgrade iOS SDK to [5.9.0](https://docs.sentiance.com/sdk/changelog/ios#5-9-0-29-jan-2021)
 - Upgrade Android SDK to [4.18.0](https://docs.sentiance.com/sdk/changelog/android#4-18-0-14-jan-2021)